Quote original thread message if enough time has passed.
Fixes #54.
