+++ config.go
@@ -11,7 +11,8 @@ import (
 )
 
 type slack struct {
-	Token string
+	Token                  string
+	ThreadQuoteIntervalSec int
 }
 
 // Config holds configuration data for Tanya
