+++ go.mod
@@ -4,8 +4,8 @@ go 1.13
 
 require (
 	github.com/BurntSushi/toml v0.3.1
-	github.com/slack-go/slack v0.8.1
+	github.com/hashicorp/golang-lru v0.5.4
+	github.com/pkg/errors v0.9.1 // indirect
+	github.com/slack-go/slack v0.8.2
 	golang.org/x/term v0.0.0-20210220032956-6a3ed077a48d
 )
-
-replace github.com/slack-go/slack => github.com/nolanlum/slack v0.2.1-0.20210225003417-b0474e47058a
